cloner/requirements.txt
trafilatura>=1.12.0
markdownify>=0.13.0
httpx>=0.27.0
Pillow>=10.0.0
python-slugify>=8.0.0
PyYAML>=6.0.1
rich>=13.0.0
typer>=0.12.0
lxml>=5.0.0
Jinja2>=3.1.0
